fix: move parseTriggerService to star-parser to fix react-refresh lint
Move parseTriggerService from saga-flow.tsx to star-parser.ts so
saga-flow.tsx only exports React components, satisfying the
react-refresh/only-export-components lint rule. Also remove unused
detectSagas function from use-pattern-files.ts.
